extract all shared/reusable code from remotehosts.py and create endpoints.py containing it
