+++ endpoints/remotehosts/remotehosts.py
@@ -6,10 +6,8 @@ Endpoint to run 1 or more engines on 1 or more remotehost systems
 
 import argparse
 import base64
-import calendar
 import copy
 from fabric import Connection
-from invoke import run
 import jsonschema
 import logging
 import os
@@ -22,6 +20,12 @@ import tempfile
 import threading
 import time
 import traceback
+
+script_path = os.path.abspath(__file__)
+script_dir = os.path.dirname(script_path)
+endpoints_dir = script_dir + "/../"
+sys.path.append(endpoints_dir)
+import endpoints
 
 TOOLBOX_HOME = os.environ.get('TOOLBOX_HOME')
 if TOOLBOX_HOME is None:
@@ -37,20 +41,6 @@ else:
     sys.path.append(str(p))
 from toolbox.json import *
 
-ROADBLOCK_HOME = os.environ.get('ROADBLOCK_HOME')
-if ROADBLOCK_HOME is None:
-    print("This script requires libraries that are provided by the roadblock project.")
-    print("Roadblock can be acquired from https://github.com/perftool-incubator/roadblock and")
-    print("then use 'export ROADBLOCK_HOME=/path/to/roadblock so that it can be located.")
-    exit(1)
-else:
-    p = Path(ROADBLOCK_HOME) / 'roadblock.py'
-    if not p.exists() or not p.is_file():
-        print("ERROR: <ROADBLOCK_HOME>/roadblock.py ('%s') does not exist!" % (p))
-        exit(2)
-    sys.path.append(str(Path(ROADBLOCK_HOME)))
-from roadblock import roadblock
-
 endpoint_defaults = {
     "controller-ip-address": None,
     "cpu-partitioning": False,
@@ -63,125 +53,6 @@ endpoint_defaults = {
     "user": "root",
     "maximum-worker-threads-count": 250
 }
-
-roadblock_exits = {
-    "success": 0,
-    "timeout": 3,
-    "abort": 4,
-    "input": 2
-}
-
-def get_result_log_level(result):
-    """
-    Determine the log level to return based on the provided result's return code.
-
-    Args:
-        result (Fabric/Invoke result): From a Fabric/Invoke run command
-
-    Globals:
-        None
-
-    Returns:
-        str: The appropriate log level to use to describe the result
-    """
-    if result.exited != 0:
-        return "error"
-    else:
-        return "info"
-
-def remote_connection(host, user):
-    """
-    Create a Fabric connection and open it
-
-    Args:
-       host (str): The IP address or hostname to connect to using Fabric
-       user (str): The username to connect as
-
-    Globals:
-        None
-
-    Returns:
-       an open Fabric Connection
-    """
-    connection = Connection(host = host, user = user)
-    attempts = 5
-    attempt = 0
-    while attempt < attempts:
-        try:
-            attempt += 1
-            connection.open()
-            if attempt > 1:
-                msg = "Connected to remote '%s' as user '%s' after %d attempts" % (host, user, attempt)
-                if args.validate:
-                    validate_comment(msg)
-                else:
-                    log.info(msg)
-            break
-        except (ssh_exception.AuthenticationException, ssh_exception.NoValidConnectionsError) as e:
-            msg = "Failed to connect to remote '%s' as user '%s' on attempt %d due to '%s'" % (host, user, attempt, str(e))
-            if args.validate:
-                validate_comment(msg)
-            else:
-                log.warning(msg)
-
-            if attempt == attempts:
-                msg = "Failed to connect to remote '%s' as user '%s' and maximum number of attempts (%d) has been exceeded.  Reraising exception '%s'" % (host, user, attempts, str(e))
-                if args.validate:
-                    validate_error(msg)
-                else:
-                    log.error(msg)
-
-                    raise e
-            else:
-                time.sleep(1)
-    return connection
-
-def run_remote(connection, command):
-    """
-    Run a command on a remote server using an existing Fabric connection
-
-    Args:
-        connection (Fabric Connection): The connection to use to run the command remotely
-        command (str): The command to run
-
-    Globals:
-        args (namespace): the script's CLI parameters
-        log: a logger instance
-
-    Returns:
-        a Fabric run result
-    """
-    debug_msg = "on remote '%s' as '%s' running command '%s'" % (connection.host, connection.user, command)
-    if args.validate:
-        if args.log_level == "debug":
-            validate_debug(debug_msg)
-    else:
-        log.debug(debug_msg, stacklevel = 2)
-
-    return connection.run(command, hide = True, warn = True)
-
-def run_local(command):
-    """
-    Run a command on the local machine using Invoke
-
-    Args:
-        command (str): The command to run on the local system
-
-    Globals:
-        args (namespace): the script's CLI parameters
-        log: a logger instance
-
-    Returns:
-        an Invoke run result
-    """
-    debug_msg = "running local command '%s'" % (command)
-    if args.validate:
-        if args.log_level == "debug":
-            validate_debug(debug_msg)
-    else:
-        log.debug(debug_msg, stacklevel = 2)
-
-    return run(command, hide = True, warn = True)
 
 def process_options():
     """
@@ -305,84 +176,6 @@ def process_options():
     
     return args
 
-def validate_log(msg):
-    """
-    Log a validation message
-
-    Args:
-        msg (str): The message to log
-
-    Globals:
-        None
-
-    Returns:
-        None
-    """
-    return print(msg)
-
-def validate_comment(msg):
-    """
-    Log a validation comment message
-
-    Args:
-        msg (str): The message to log
-
-    Globals:
-        None
-
-    Returns:
-        None
-    """
-    return print("#" + msg)
-
-def validate_error(msg):
-    """
-    Log a validation error message
-
-    Args:
-        msg (str): The message to log
-
-    Globals:
-        None
-
-    Returns:
-        None
-    """
-    return print("ERROR: " + msg)
-
-def validate_debug(msg):
-    """
-    Log a vlidation debug message
-
-    Args:
-        msg (str): The message to log
-
-    Globals:
-        None
-
-    Returns:
-        None
-    """
-    return print("# DEBUG: " + msg)
-
-def cli_stream():
-    """
-    Recreate the argument list that the script was called with
-
-    Args:
-        None
-
-    Globals:
-        sys.argv (list): The user provided CLI arguments
-
-    Returns:
-        stream (str): A formatted string of the user provided CLI arguments
-    """
-    stream = ""
-    for i in range(1, len(sys.argv)):
-        stream += " %s" % sys.argv[i]
-    return stream
-
 def validate():
     """
     Validate the input and return configuration details to the caller
@@ -396,34 +189,34 @@ def validate():
     Returns:
         int: zero for success / non-zero for failure
     """
-    validate_comment("params: %s" % (cli_stream()))
+    endpoints.validate_comment("params: %s" % (endpoints.cli_stream()))
 
-    validate_comment("argparse: %s" % (args))
+    endpoints.validate_comment("argparse: %s" % (args))
 
-    validate_comment("run-file: %s" % (args.run_file))
-    validate_comment("endpoint-index: %d" % (args.endpoint_index))
+    endpoints.validate_comment("run-file: %s" % (args.run_file))
+    endpoints.validate_comment("endpoint-index: %d" % (args.endpoint_index))
 
     json, err = load_json_file(args.run_file)
     if json is None:
-        validate_error(err)
+        endpoints.validate_error(err)
         return 1
-    validate_comment("run-file: %s" % (json))
+    endpoints.validate_comment("run-file: %s" % (json))
 
     valid, err = validate_schema(json, args.rickshaw_dir + "/util/JSON/schema.json")
     if not valid:
-        validate_error(err)
+        endpoints.validate_error(err)
         return 1
 
     if args.endpoint_index >= len(json["endpoints"]):
-        validate_error("endpoint_index %d does not exist in endpoints array from run-file" % (args.endpoint_index))
+        endpoints.validate_error("endpoint_index %d does not exist in endpoints array from run-file" % (args.endpoint_index))
         return 1
 
     endpoint_settings = json["endpoints"][args.endpoint_index]
-    validate_comment("endpoint-settings: %s" % (endpoint_settings))
+    endpoints.validate_comment("endpoint-settings: %s" % (endpoint_settings))
 
     valid, err = validate_schema(endpoint_settings, args.rickshaw_dir + "/schema/remotehosts.json")
     if not valid:
-        validate_error(err)
+        endpoints.validate_error(err)
         return 1
 
     for endpoint_idx,endpoint in enumerate(json["endpoints"]):
@@ -431,21 +224,21 @@ def validate():
             continue
 
         if json["endpoints"][endpoint_idx]["type"] == "remotehosts":
-            validate_error("You can only specify one instance of the remotehosts endpoint.  If there is a reason you need multiple remotehosts endpoints then it should be seen as a bug in the remotehosts endpoint.")
+            endpoints.validate_error("You can only specify one instance of the remotehosts endpoint.  If there is a reason you need multiple remotehosts endpoints then it should be seen as a bug in the remotehosts endpoint.")
 
     rickshaw_settings, err = load_json_file(args.base_run_dir + "/config/rickshaw-settings.json.xz", uselzma = True)
     if rickshaw_settings is None:
-        validate_error(err)
+        endpoints.validate_error(err)
         return 1
-    validate_comment("rickshaw-settings: %s" % rickshaw_settings)
+    endpoints.validate_comment("rickshaw-settings: %s" % rickshaw_settings)
 
     endpoint_settings = normalize_endpoint_settings(endpoint_settings, rickshaw_settings)
     if endpoint_settings is None:
         return 1
-    validate_comment("normalized endpoint-settings: %s" % (endpoint_settings))
+    endpoints.validate_comment("normalized endpoint-settings: %s" % (endpoint_settings))
 
-    benchmark_engine_mapping = build_benchmark_engine_mapping(json["benchmarks"])
-    validate_comment("benchmark-engine-mapping: %s" % (benchmark_engine_mapping))
+    benchmark_engine_mapping = endpoints.build_benchmark_engine_mapping(json["benchmarks"])
+    endpoints.validate_comment("benchmark-engine-mapping: %s" % (benchmark_engine_mapping))
 
     engines = dict()
     userenvs = []
@@ -460,11 +253,11 @@ def validate():
         if not remote["config"]["settings"]["userenv"] in userenvs:
             userenvs.append(remote["config"]["settings"]["userenv"])
 
-    validate_comment("engines: %s" % (engines))
+    endpoints.validate_comment("engines: %s" % (engines))
     for role in engines.keys():
-        validate_log("%s %s" % (role, " ".join(map(str, engines[role]))))
+        endpoints.validate_log("%s %s" % (role, " ".join(map(str, engines[role]))))
         if len(engines[role]) != len(set(engines[role])):
-            validate_error("There are duplicate IDs present for %s" % (role))
+            endpoints.validate_error("There are duplicate IDs present for %s" % (role))
         for engine_id in engines[role]:
             found_engine = False
             for benchmark in benchmark_engine_mapping.keys():
@@ -472,11 +265,11 @@ def validate():
                     found_engine = True
                     break
             if not found_engine:
-                validate_error("Could not find a benchmark mapping for engine ID %d" % (engine_id))
+                endpoints.validate_error("Could not find a benchmark mapping for engine ID %d" % (engine_id))
 
-    validate_comment("userenvs: %s" % (userenvs))
+    endpoints.validate_comment("userenvs: %s" % (userenvs))
     for userenv in userenvs:
-        validate_log("userenv %s" % (userenv))
+        endpoints.validate_log("userenv %s" % (userenv))
 
     remotes = dict()
     for remote in endpoint_settings["remotes"]:
@@ -484,53 +277,31 @@ def validate():
             remotes[remote["config"]["host"]] = dict()
         if not remote["config"]["settings"]["remote-user"] in remotes[remote["config"]["host"]]:
             remotes[remote["config"]["host"]][remote["config"]["settings"]["remote-user"]] = True
-    validate_comment("remotes: %s" % (remotes))
+    endpoints.validate_comment("remotes: %s" % (remotes))
 
+    debug_output = False
+    if args.log_level == "debug":
+        debug_output = True
     for remote in remotes.keys():
         for remote_user in remotes[remote].keys():
             try:
-                with remote_connection(remote, remote_user) as c:
-                    result = run_remote(c, "uptime")
-                    validate_comment("remote login verification for %s with user %s: rc=%d and stdout=[%s] annd stderr=[%s]" % (remote, remote_user, result.exited, result.stdout.rstrip('\n'), result.stderr.rstrip('\n')))
+                with endpoints.remote_connection(remote, remote_user, validate = True) as c:
+                    result = endpoints.run_remote(c, "uptime", validate = True, debug = debug_output)
+                    endpoints.validate_comment("remote login verification for %s with user %s: rc=%d and stdout=[%s] annd stderr=[%s]" % (remote, remote_user, result.exited, result.stdout.rstrip('\n'), result.stderr.rstrip('\n')))
 
-                    result = run_remote(c, "podman --version")
-                    validate_comment("remote podman presence check for %s: rc=%d and stdout=[%s] and stderr=[%s]" % (remote, result.exited, result.stdout.rstrip('\n'), result.stderr.rstrip('\n')))
+                    result = endpoints.run_remote(c, "podman --version", validate = True, debug = debug_output)
+                    endpoints.validate_comment("remote podman presence check for %s: rc=%d and stdout=[%s] and stderr=[%s]" % (remote, result.exited, result.stdout.rstrip('\n'), result.stderr.rstrip('\n')))
                     if result.exited != 0:
-                        result = run_remote(c, "yum install -y podman")
-                        validate_comment("remote podman installation for %s: rc=%d" % (remote, result.exited))
+                        result = endpoints.run_remote(c, "yum install -y podman", validate = True, debug = debug_output)
+                        endpoints.validate_comment("remote podman installation for %s: rc=%d" % (remote, result.exited))
                         if result.exited != 0:
-                            validate_error("Could not install podman to remote %s" % (remote))
-                            validate_error("stdout:\n%s" % (result.stdout))
-                            validate_error("stderr:\n%s" % (result.stderr))
+                            endpoints.validate_error("Could not install podman to remote %s" % (remote))
+                            endpoints.validate_error("stdout:\n%s" % (result.stdout))
+                            endpoints.validate_error("stderr:\n%s" % (result.stderr))
             except ssh_exception.AuthenticationException as e:
-                validate_error("remote login verification for %s with user %s resulted in an authentication exception '%s'" % (remote, remote_user, str(e)))
+                endpoints.validate_error("remote login verification for %s with user %s resulted in an authentication exception '%s'" % (remote, remote_user, str(e)))
             except ssh_exception.NoValidConnectionsError as e:
-                validate_error("remote login verification for %s with user %s resulted in an connection exception '%s'" % (remote, remote_user, str(e)))
-
-    return 0
-
-def log_cli():
-    """
-    Log the script invocation details in a readable form
-
-    Args:
-        None
-
-    Globals:
-        args (namespace): the script's CLI parameters
-        log: a logger instance
-
-    Returns:
-        0
-    """
-    log.info("Logging CLI")
-
-    log.info("CLI parameters:\n%s" % (cli_stream()))
-
-    the_args = dict()
-    for arg in args.__dict__:
-        the_args[str(arg)] = args.__dict__[arg]
-    log.info("argparse:\n %s" % (dump_json(the_args)))
+                endpoints.validate_error("remote login verification for %s with user %s resulted in an connection exception '%s'" % (remote, remote_user, str(e)))
 
     return 0
 
@@ -611,70 +382,21 @@ def log_settings(mode = "all"):
     """
     match mode:
         case "benchmark-mapping":
-            return log.info("settings[benchmark-mapping]:\n%s" % (dump_json(settings["engines"]["benchmark-mapping"])), stacklevel = 2)
+            return log.info("settings[benchmark-mapping]:\n%s" % (endpoints.dump_json(settings["engines"]["benchmark-mapping"])), stacklevel = 2)
         case "engines":
-            return log.info("settings[engines]:\n%s" % (dump_json(settings["engines"])), stacklevel = 2)
+            return log.info("settings[engines]:\n%s" % (endpoints.dump_json(settings["engines"])), stacklevel = 2)
         case "misc":
-            return log.info("settings[misc]:\n%s" % (dump_json(settings["misc"])), stacklevel = 2)
+            return log.info("settings[misc]:\n%s" % (endpoints.dump_json(settings["misc"])), stacklevel = 2)
         case "dirs":
-            return log.info("settings[dirs]:\n%s" % (dump_json(settings["dirs"])), stacklevel = 2)
+            return log.info("settings[dirs]:\n%s" % (endpoints.dump_json(settings["dirs"])), stacklevel = 2)
         case "endpoint":
-            return log.info("settings[endpoint]:\n%s" % (dump_json(settings["run-file"]["endpoints"][args.endpoint_index])), stacklevel = 2)
+            return log.info("settings[endpoint]:\n%s" % (endpoints.dump_json(settings["run-file"]["endpoints"][args.endpoint_index])), stacklevel = 2)
         case "rickshaw":
-            return log.info("settings[rickshaw]:\n%s" % (dump_json(settings["rickshaw"])), stacklevel = 2)
+            return log.info("settings[rickshaw]:\n%s" % (endpoints.dump_json(settings["rickshaw"])), stacklevel = 2)
         case "run-file":
-            return log.info("settings[run-file]:\n%s" % (dump_json(settings["run-file"])), stacklevel = 2)
+            return log.info("settings[run-file]:\n%s" % (endpoints.dump_json(settings["run-file"])), stacklevel = 2)
         case "all" | _:
-            return log.info("settings:\n%s" % (dump_json(settings)), stacklevel = 2)
-
-def not_json_serializable(obj):
-    """
-    Convert non-serializable variables into something that can be handled by the JSON conversion process'
-
-    Args:
-        obj: a non-serializable variable
-
-    Globals:
-        None
-
-    Returns:
-        str: a representation of the object that can be serialized into a json property value
-    """
-    try:
-        return(obj.to_dictionary())
-    except AttributeError:
-        return(repr(obj))
-
-def dump_json(obj):
-    """
-    Convert a variable into a formatted JSON string
-
-    Args:
-        obj: A variable of potentially many types to convert into a JSON string
-
-    Globals:
-        None
-
-    Returns:
-        str: A formatted string containing the JSON representation of obj
-    """
-    return json.dumps(obj, indent = 4, separators=(',', ': '), sort_keys = True, default = not_json_serializable)
-
-def my_make_dirs(mydir):
-    """
-    Created the requested directory (recursively if necessary)
-
-    Args:
-        mydir (str): The directory that needs to be created
-
-    Globals:
-        log: a logger instance
-
-    Returns:
-        None
-    """
-    log.info("Creating directory %s (recurisvely if necessary)" % (mydir), stacklevel = 2)
-    return os.makedirs(mydir, exist_ok = True)
+            return log.info("settings:\n%s" % (endpoints.dump_json(settings)), stacklevel = 2)
 
 def create_local_dirs():
     """
@@ -691,10 +413,10 @@ def create_local_dirs():
         0
     """
     log.info("Creating local directories")
-    my_make_dirs(settings["dirs"]["local"]["run"])
-    my_make_dirs(settings["dirs"]["local"]["engine-logs"])
-    my_make_dirs(settings["dirs"]["local"]["roadblock-msgs"])
-    my_make_dirs(settings["dirs"]["local"]["sysinfo"])
+    endpoints.my_make_dirs(settings["dirs"]["local"]["run"])
+    endpoints.my_make_dirs(settings["dirs"]["local"]["engine-logs"])
+    endpoints.my_make_dirs(settings["dirs"]["local"]["roadblock-msgs"])
+    endpoints.my_make_dirs(settings["dirs"]["local"]["sysinfo"])
     return 0
 
 def load_settings():
@@ -753,7 +475,7 @@ def load_settings():
     log.info("Building benchmark engine mapping")
     if not "engines" in settings:
         settings["engines"] = dict()
-    settings["engines"]["benchmark-mapping"] = build_benchmark_engine_mapping(settings["run-file"]["benchmarks"])
+    settings["engines"]["benchmark-mapping"] = endpoints.build_benchmark_engine_mapping(settings["run-file"]["benchmarks"])
     log_settings(mode = "benchmark-mapping")
 
     log.info("Loading SSH private key into misc settings")
@@ -770,185 +492,6 @@ def load_settings():
     log_settings(mode = "misc")
 
     return 0
-
-def expand_id_range(id_range):
-    """
-    Helper function to expand an ID range
-
-    Args:
-        id_range (str): An ID range in one of the following forms [ "A-E", "F" ]
-
-    Globals:
-        None
-
-    Returns:
-        expanded_ids (list): A list containing the broken out IDs in integer format such as [ A, B, C, D, E ] or [ F ]
-    """
-    expanded_ids = []
-
-    split_id_range = id_range.split("-")
-    if len(split_id_range) == 1:
-        expanded_ids.append(int(split_id_range[0]))
-    else:
-        split_id_range = list(map(int, split_id_range))
-        if split_id_range[0] >= 0 and split_id_range[0] < split_id_range[1]:
-            for subid in range(split_id_range[0], split_id_range[1]+1):
-                expanded_ids.append(subid)
-        else:
-            raise ValueError("Invalid id range: %s" % (id_range))
-
-    return expanded_ids
-
-def expand_ids(ids):
-    """
-    Take a IDs specification and expand it
-
-    Args:
-        ids (str, int, or list): A complex variable containing one of many forms of ID representation
-                                 such as [ "A-E", "A-E+H-K+M", "A", A, [ A, "C-E", "H" ] ]
-
-    Globals:
-        None
-
-    Returns:
-        new_ids (list): A list containing the broken out IDs in integer format such as [ A, B, C, D, E] or
-                        [ A, B, C, D, E, H, J, K, M ] or [ A ] or [ A ] or [ A, C, D, E, H ]
-    """
-    new_ids = []
-
-    if isinstance(ids, str):
-        subids = ids.split("+")
-        for subid in subids:
-            new_ids.extend(expand_id_range(subid))
-    elif isinstance(ids, int):
-        new_ids.append(ids)
-    elif isinstance(ids, list):
-        for id in ids:
-            if isinstance(id, int):
-                new_ids.append(id)
-            elif isinstance(id, str):
-                new_ids.extend(expand_id_range(id))
-
-    new_ids.sort()
-
-    return new_ids
-
-def is_ip(ip_address):
-    """
-    Determine if the provided variable contains an IPv4 of IPv6 address
-
-    Args:
-        ip_address (str): A string to check if it is a valid IPv4 or IPv5 address (it could be a hostname)
-
-    Globals:
-        None
-
-    Returns:
-        True or False
-    """
-    msg = "Checking to see if '%s' is an IP address" % (ip_address)
-    if args.validate:
-        validate_comment(msg)
-    else:
-        log.info(msg)
-
-    # check for IPv4
-    m = re.search(r"[1-9][0-9]{0,2}\.[1-9][0-9]{0,2}\.[1-9][0-9]{0,2}\.[1-9][0-9]{0,2}", ip_address)
-    if m:
-        return True
-
-    # check for IPv6
-    m = re.search(r"[0-9a-fA-F]{1,4}:[0-9a-fA-F]{1,4}:[0-9a-fA-F]{1,4}:[0-9a-fA-F]{1,4}:[0-9a-fA-F]{1,4}:[0-9a-fA-F]{1,4}:[0-9a-fA-F]{1,4}:[0-9a-fA-F]", ip_address)
-    if m:
-        return True
-
-    return False
-
-def get_controller_ip(host):
-    """
-    Determine the correct controller IP address for the provided host to contact the controller at
-
-    Args:
-        host (str): Either a hostname or an IP address for a remote
-
-    Globals:
-        None
-
-    Returns:
-        controller_ip (str): The controller's IP address that the specified remote can use to contact it
-    """
-    controller_ip = None
-
-    if host == "localhost":
-        cmd = "ip addr show lo"
-        result = run_local(cmd)
-        if result.exited != 0:
-            raise ValueError("Failed to successfully run '%s'" % (cmd))
-        for line in result.stdout.splitlines():
-            # looking for something like the following line
-            #    inet 127.0.0.1/8 scope host lo
-            m = re.search(r"int\s", line)
-            if m:
-                split_line = line.split(" ")
-                if len(split_line) > 1:
-                    split_line = split_line[1]
-                    if len(split_line) > 1:
-                        controller_ip = split_line[0]
-                        break
-        if controller_ip is None or not is_ip(controller_ip):
-            raise ValueError("Failed to map localhost to loopback IP address (got '%s')" % controller_ip)
-        else:
-            host = controller_ip
-            controller_ip = None
-
-    remote_ip = None
-    if is_ip(host):
-        cmd = "ip addr"
-        result = run_local(cmd)
-        if result.exited != 0:
-            raise ValueError("Failed to successfully run '%s'" % (cmd))
-        for line in result.stdout.splitlines():
-            m = re.search(host, line)
-            if m:
-                # the provided ip address is mine so it is the one to use
-                return host
-        # the provided ip address is not mine -- so it is the remotes's
-        remote_ip = host
-    else:
-        cmd = "host %s" % (host)
-        result = run_local(cmd)
-        if result.exited != 0:
-            raise ValueError("Failed to successfully run '%s'" % (cmd))
-        for line in result.stdout.splitlines():
-            m = re.search(r"has address", line)
-            if m:
-                split_line = line.split("has address ")
-                if len(split_line) > 1:
-                    remote_ip = split_line[1]
-                    break
-        if remote_ip is None or not is_ip(remote_ip):
-            raise ValueError("Failed to determine remote IP address (got '%s')" % (endpoint_ip))
-
-    # now that we have the remote's ip address, figure out what
-    # controller ip this remote will need to use to contact the
-    # controller
-    cmd = "ip route get %s" % (remote_ip)
-    result = run_local(cmd)
-    if result.exited != 0:
-        raise ValueError("Failed to successfully run '%s'" % (cmd))
-    for line in result.stdout.splitlines():
-        m = re.search(r"src ", line)
-        if m:
-            split_line = line.split("src ")
-            if len(split_line) > 1:
-                split_line = split_line[1].split(" ")
-                if len(split_line) > 1:
-                    controller_ip = split_line[0]
-                    break
-    if controller_ip is None or not is_ip(controller_ip):
-        raise ValueError("Failed to determine controller IP address (got '%s')" % (controller_ip))
-
-    return controller_ip
 
 def normalize_endpoint_settings(endpoint, rickshaw):
     """
@@ -998,12 +541,12 @@ def normalize_endpoint_settings(endpoint, rickshaw):
                 remote["config"]["settings"]["controller-ip-address"] = cached_controller_ips[remote["config"]["host"]]
             else:
                 try:
-                    remote["config"]["settings"]["controller-ip-address"] = get_controller_ip(remote["config"]["host"])
+                    remote["config"]["settings"]["controller-ip-address"] = endpoints.get_controller_ip(remote["config"]["host"])
                     cached_controller_ips[remote["config"]["host"]] = remote["config"]["settings"]["controller-ip-address"]
                 except ValueError as e:
                     msg = "While determining controller IP address for remote '%s' encountered exception '%s'" % (remote["config"]["host"], str(e))
                     if args.validate:
-                        validate_error(msg)
+                        endpoints.validate_error(msg)
                     else:
                         log.error(msg)
                     return None
@@ -1012,11 +555,11 @@ def normalize_endpoint_settings(endpoint, rickshaw):
             if engine["role"] == "profiler":
                 continue
             try:
-                engine["ids"] = expand_ids(engine["ids"])
+                engine["ids"] = endpoints.expand_ids(engine["ids"])
             except ValueError as e:
                 msg = "While expanding '%s' encountered exception '%s'" % (engine["ids"], str(e))
                 if args.validate:
-                    validate_error(msg)
+                    endpoints.validate_error(msg)
                 else:
                     log.error(msg)
                 return None
@@ -1194,30 +737,6 @@ def build_unique_remote_configs():
 
     return 0
 
-def build_benchmark_engine_mapping(benchmarks):
-    """
-    Build the benchmark-to-engine mapping
-
-    Args:
-        benchmarks (list): The list of benchmarks from the run-file
-
-    Globals:
-        None
-
-    Returns:
-        mapping (dict): A collection of dictionaries which maps particular client/server IDs to a particular benchmark
-    """
-    mapping = dict()
-
-    for benchmark_idx,benchmark in enumerate(benchmarks):
-        benchmark_id = benchmark["name"] + "-" + str(benchmark_idx)
-        mapping[benchmark_id] = {
-            "name": benchmark["name"],
-            "ids": expand_ids(benchmark["ids"])
-        }
-
-    return mapping
-
 def get_profiler(profiler_id):
     """
     Get the profiler that a specific profiler engine should be running from the profiler mapping
@@ -1323,16 +842,16 @@ def image_pull_worker_thread(thread_id, work_queue, threads_rcs):
 
         thread_logger(thread_name, "Remote user is %s" % (my_run_file_remote["config"]["settings"]["remote-user"]), remote_name = remote)
 
-        with remote_connection(remote, my_run_file_remote["config"]["settings"]["remote-user"]) as c:
+        with endpoints.remote_connection(remote, my_run_file_remote["config"]["settings"]["remote-user"]) as c:
             for image in my_unique_remote["images"]:
-                result = run_remote(c, "podman pull " + image)
+                result = endpoints.run_remote(c, "podman pull " + image)
                 loglevel = "info"
                 if result.exited != 0:
                     loglevel = "error"
                 thread_logger(thread_name, "Attempted to pull %s with return code %d:\nstdout:\n%sstderr:\n%s" % (image, result.exited, result.stdout, result.stderr), log_level = loglevel, remote_name = remote)
                 rc += result.exited
 
-                result = run_remote(c, "echo '" + image + " " + str(int(time.time())) + " " + args.run_id + "' >> " + settings["dirs"]["remote"]["base"] + "/remotehosts-container-image-census")
+                result = endpoints.run_remote(c, "echo '" + image + " " + str(int(time.time())) + " " + args.run_id + "' >> " + settings["dirs"]["remote"]["base"] + "/remotehosts-container-image-census")
                 loglevel = "info"
                 if result.exited != 0:
                     loglevel = "error"
@@ -1473,7 +992,7 @@ def create_thread_pool(description, acronym, work, worker_threads_count, worker_
         else:
             thread_logger("MAIN", "Skipping join of thread %s because it does not exist" % (thread_name), log_level = "warning")
 
-    thread_logger("MAIN", "Return codes for each %s:\n%s" % (acronym, dump_json(worker_threads_rcs)))
+    thread_logger("MAIN", "Return codes for each %s:\n%s" % (acronym, endpoints.dump_json(worker_threads_rcs)))
 
     return
 
@@ -1556,10 +1075,10 @@ def remote_mkdirs_worker_thread(thread_id, work_queue, threads_rcs):
         my_unique_remote = settings["engines"]["remotes"][remote]
         my_run_file_remote = settings["run-file"]["endpoints"][args.endpoint_index]["remotes"][my_unique_remote["run-file-idx"][0]]
 
-        with remote_connection(remote, my_run_file_remote["config"]["settings"]["remote-user"]) as con:
+        with endpoints.remote_connection(remote, my_run_file_remote["config"]["settings"]["remote-user"]) as con:
             for remote_dir in settings["dirs"]["remote"].keys():
-                result = run_remote(con, "mkdir --parents --verbose " + settings["dirs"]["remote"][remote_dir])
-                thread_logger(thread_name, "Remote attempted to mkdir %s with return code %d:\nstdout:\n%sstderr:\n%s" % (settings["dirs"]["remote"][remote_dir], result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote)
+                result = endpoints.run_remote(con, "mkdir --parents --verbose " + settings["dirs"]["remote"][remote_dir])
+                thread_logger(thread_name, "Remote attempted to mkdir %s with return code %d:\nstdout:\n%sstderr:\n%s" % (settings["dirs"]["remote"][remote_dir], result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote)
                 rc += result.exited
 
         thread_logger(thread_name, "Notifying work queue that job processing is complete", remote_name = remote)
@@ -1714,10 +1233,10 @@ def create_podman(thread_name, remote_name, engine_name, container_name, connect
 
     create_cmd.append(image)
 
-    thread_logger(thread_name, "Podman create command is:\n%s" % (dump_json(create_cmd)), remote_name = remote_name, engine_name = engine_name)
+    thread_logger(thread_name, "Podman create command is:\n%s" % (endpoints.dump_json(create_cmd)), remote_name = remote_name, engine_name = engine_name)
 
-    result = run_remote(connection, " ".join(create_cmd))
-    thread_logger(thread_name, "Creating container with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+    result = endpoints.run_remote(connection, " ".join(create_cmd))
+    thread_logger(thread_name, "Creating container with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
     return
 
@@ -1757,10 +1276,10 @@ def create_chroot(thread_name, remote_name, engine_name, container_name, connect
                    "--name=" + container_name,
                    image ]
 
-    thread_logger(thread_name, "Podman create command is:\n%s" % (dump_json(create_cmd)), remote_name = remote_name, engine_name = engine_name)
+    thread_logger(thread_name, "Podman create command is:\n%s" % (endpoints.dump_json(create_cmd)), remote_name = remote_name, engine_name = engine_name)
 
-    result = run_remote(connection, " ".join(create_cmd))
-    thread_logger(thread_name, "Creating container with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+    result = endpoints.run_remote(connection, " ".join(create_cmd))
+    thread_logger(thread_name, "Creating container with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
     if result.exited == 0:
         create_info["id"] = result.stdout.rstrip('\n')
     else:
@@ -1770,8 +1289,8 @@ def create_chroot(thread_name, remote_name, engine_name, container_name, connect
                   "mount",
                   container_name ]
 
-    result = run_remote(connection, " ".join(mount_cmd))
-    thread_logger(thread_name, "Mounting container with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+    result = endpoints.run_remote(connection, " ".join(mount_cmd))
+    thread_logger(thread_name, "Mounting container with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
     if result.exited == 0:
         create_info["mount"] = result.stdout.rstrip('\n')
     else:
@@ -1828,22 +1347,22 @@ def create_chroot(thread_name, remote_name, engine_name, container_name, connect
 
         mount["dest"] = create_info["mount"] + mount["dest"]
 
-        thread_logger(thread_name, "Procesing mount:\n%s" % (dump_json(mount)), remote_name = remote_name, engine_name = engine_name)
+        thread_logger(thread_name, "Procesing mount:\n%s" % (endpoints.dump_json(mount)), remote_name = remote_name, engine_name = engine_name)
 
-        result = run_remote(connection, "mkdir --parents --verbose " + mount["dest"])
-        thread_logger(thread_name, "Creating '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (mount["dest"], result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+        result = endpoints.run_remote(connection, "mkdir --parents --verbose " + mount["dest"])
+        thread_logger(thread_name, "Creating '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (mount["dest"], result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
         if mount["rbind"]:
-            result = run_remote(connection, "mount --verbose --options rbind " + mount["src"] + " " + mount["dest"])
-            thread_logger(thread_name, "rbind mounting '%s' to '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (mount["src"], mount["dest"], result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+            result = endpoints.run_remote(connection, "mount --verbose --options rbind " + mount["src"] + " " + mount["dest"])
+            thread_logger(thread_name, "rbind mounting '%s' to '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (mount["src"], mount["dest"], result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
-            result = run_remote(connection, "mount --verbose --make-rslave " + mount["dest"])
-            thread_logger(thread_name, "making rslave '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (mount["dest"], result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+            result = endpoints.run_remote(connection, "mount --verbose --make-rslave " + mount["dest"])
+            thread_logger(thread_name, "making rslave '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (mount["dest"], result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
             create_info["mounts"]["rbind"].append(mount["dest"])
         else:
-            result = run_remote(connection, "mount --verbose --options bind " + mount["src"] + " " + mount["dest"])
-            thread_logger(thread_name, "bind mounting '%s' to '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (mount["src"], mount["dest"], result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+            result = endpoints.run_remote(connection, "mount --verbose --options bind " + mount["src"] + " " + mount["dest"])
+            thread_logger(thread_name, "bind mounting '%s' to '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (mount["src"], mount["dest"], result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
             create_info["mounts"]["regular"].append(mount["dest"])
 
@@ -1855,10 +1374,10 @@ def create_chroot(thread_name, remote_name, engine_name, container_name, connect
     for etc_file in [ "hosts", "resolv.conf" ]:
         src_etc_file = "/etc/" + etc_file
         dst_etc_dir = create_info["mount"] + "/etc"
-        result = run_remote(connection, "cp --verbose " + src_etc_file + " " + dst_etc_dir)
-        thread_logger(thread_name, "Remotely copied %s to %s" % (src_etc_file, dst_etc_dir), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+        result = endpoints.run_remote(connection, "cp --verbose " + src_etc_file + " " + dst_etc_dir)
+        thread_logger(thread_name, "Remotely copied %s to %s" % (src_etc_file, dst_etc_dir), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
-    thread_logger(thread_name, "chroot create info:\n%s" % (dump_json(create_info)), remote_name = remote_name, engine_name = engine_name)
+    thread_logger(thread_name, "chroot create info:\n%s" % (endpoints.dump_json(create_info)), remote_name = remote_name, engine_name = engine_name)
 
     return create_info
 
@@ -1887,8 +1406,8 @@ def start_podman(thread_name, remote_name, engine_name, container_name, connecti
         container_name
     ]
 
-    result = run_remote(connection, " ".join(start_cmd))
-    thread_logger(thread_name, "Starting container with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+    result = endpoints.run_remote(connection, " ".join(start_cmd))
+    thread_logger(thread_name, "Starting container with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
     return
 
@@ -1963,10 +1482,10 @@ def start_chroot(thread_name, remote_name, engine_name, container_name, connecti
         "&"
     ])
 
-    thread_logger(thread_name, "chroot start command is:\n%s" % (dump_json(start_cmd)), remote_name = remote_name, engine_name = engine_name)
+    thread_logger(thread_name, "chroot start command is:\n%s" % (endpoints.dump_json(start_cmd)), remote_name = remote_name, engine_name = engine_name)
 
-    result = run_remote(connection, " ".join(start_cmd))
-    thread_logger(thread_name, "Starting chroot with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+    result = endpoints.run_remote(connection, " ".join(start_cmd))
+    thread_logger(thread_name, "Starting chroot with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
     return
 
@@ -2012,7 +1531,7 @@ def launch_engines_worker_thread(thread_id, work_queue, threads_rcs):
         thread_logger(thread_name, "Processing remote '%s' at index %d" % (remote["config"]["host"], remote_idx), remote_name = remote_name)
         thread_logger(thread_name, "Remote user is %s" % (remote["config"]["settings"]["remote-user"]), remote_name = remote_name)
 
-        with remote_connection(remote["config"]["host"], remote["config"]["settings"]["remote-user"]) as con:
+        with endpoints.remote_connection(remote["config"]["host"], remote["config"]["settings"]["remote-user"]) as con:
             for engine in remote["engines"]:
                 for engine_id in engine["ids"]:
                     engine_name = "%s-%s" % (engine["role"], str(engine_id))
@@ -2036,14 +1555,14 @@ def launch_engines_worker_thread(thread_id, work_queue, threads_rcs):
 
                     thread_logger(thread_name, "host-mounts is %s" % (remote["config"]["settings"]["host-mounts"]), remote_name = remote_name, engine_name = engine_name)
 
-                    result = run_remote(con, "podman ps --all --filter 'name=" + container_name + "' --format '{{.Names}}'")
+                    result = endpoints.run_remote(con, "podman ps --all --filter 'name=" + container_name + "' --format '{{.Names}}'")
                     thread_logger(thread_name, "Check for existing container with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), remote_name = remote_name, engine_name = engine_name)
                     if result.exited != 0:
                         thread_logger(thread_name, "Check for existing container exited with non-zero return code %d" % (result.exited), log_level = "error", remote_name = remote_name, engine_name = engine_name)
                     if result.stdout.rstrip('\n') == container_name:
                         thread_logger(thread_name, "Found existing container '%s'" % (container_name), log_level = "warning", remote_name = remote_name, engine_name = engine_name)
 
-                        result = run_remote(con, "podman rm --force " + container_name)
+                        result = endpoints.run_remote(con, "podman rm --force " + container_name)
                         thread_logger(thread_name, "Forced removal of existing container with name '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (container_name, result.exited, result.stdout, result.stderr), remote_name = remote_name, engine_name = engine_name)
                         if result.exited != 0:
                             thread_logger(thread_name, "Forced removal of existing container exited with non-zero return code %d" % (result.exited), log_level = 'error', remote_name = remote_name, engine_name = engine_name)
@@ -2138,141 +1657,6 @@ def launch_engines():
 
     return 0
 
-def create_roadblock_msg(recipient_type, recipient_id, payload_type, payload):
-    """
-    Create a user built roadblock message
-
-    Args:
-        recipient_type (str): What type of roadblock participant ("leader" or "follower" or "all") is the message for
-        recipient_id (str): What is the specific name/ID of the intended message recipient
-
-    Globals:
-        log: a logger instance
-
-    Returns:
-        msg (dict): The generated message
-    """
-    msg = [
-        {
-            "recipient": {
-                "type": recipient_type,
-                "id": recipient_id,
-            },
-            payload_type: payload
-        }
-    ]
-
-    json_msg = dump_json(msg)
-    log.info("Creating new roadblock message for recipient type '%s' with recipient id '%s':\n%s" % (recipient_type, recipient_id, json_msg), stacklevel = 2)
-
-    return msg
-
-def do_roadblock(label = None, timeout = None, messages = None, wait_for = None, abort = None):
-    """
-    Run a roadblock
-
-    Args:
-        label (str): The name of the roadblock to participate in
-        timeout (int): An optional timeout value to use for the roadblock
-        messages (str): An optional messages file to send
-        wait_for (str): An optional command to wait on to complete the roadblock
-        abort (bool): An optional parameter specifying that a abort should be sent
-
-    Globals:
-        log: a logger instance
-
-    Returns:
-        rc (int): The return code for the roadblock
-    """
-    if label is None:
-        log.error("No roadblock label specified", stacklevel = 2)
-        raise ValueError("No roadblock label specified")
-
-    log.info("Processing roadblock '%s'" % (label), stacklevel = 2)
-    uuid = "%s:%s" % (args.roadblock_id, label)
-    log.info("[%s] Roadblock uuid is '%s'" % (label, uuid))
-
-    if timeout is None:
-        timeout = 300
-        log.info("[%s] No roadblock timeout specified, defaulting to %d" % (label, timeout))
-    else:
-        log.info("[%s] Roadblock timeout set to %d" % (label, timeout))
-
-    if messages is None:
-        log.info("[%s] No roadblock messages to send" % (label))
-    else:
-        log.info("[%s] Sending roadblock messages %s" % (label, messages))
-
-    if wait_for is None:
-        log.info("[%s] No roadblock wait-for" % (label))
-    else:
-        wait_for_log = tempfile.mkstemp(suffix = "log")
-        wait_for_log[0].close()
-        wait_for_log = wait_for_log[1]
-        log.info("[%s] Going to run this wait-for command: %s" % (label, wait_For))
-        log.info("[%s] Going to log wait-for to this file: %s" % (label, wait_for_log))
-
-    if not abort is None:
-        log.info("[%s] Going to send an abort")
-
-    msgs_log_file = settings["dirs"]["local"]["roadblock-msgs"] + "/" + label + ".json"
-    log.info("[%s] Logging messages to: %s" % (label, msgs_log_file))
-
-    redis_server = "localhost"
-    leader = "controller"
-
-    result = run_local("ping -w 10 -c 4 " + redis_server)
-    ping_log_msg = "[%s] Pinged redis server '%s' with return code %d:\nstdout:\n%sstderr:\n%s" % (label, redis_server, result.exited, result.stdout, result.stderr)
-    if result.exited != 0:
-        log.error(ping_log_mesg)
-    else:
-        log.info(ping_log_msg)
-
-    attempts = 0
-    rc = -1
-    while attempts < args.max_rb_attempts and rc != roadblock_exits["success"] and rc != roadblock_exits["abort"]:
-        attempts += 1
-        log.info("[%s] Attempt number: %d" % (label, attempts))
-
-        my_roadblock = roadblock(log, False)
-        my_roadblock.set_uuid("%d:%s" % (attempts, uuid))
-        my_roadblock.set_role("follower")
-        my_roadblock.set_follower_id(args.endpoint_label)
-        my_roadblock.set_leader_id(leader)
-        my_roadblock.set_timeout(timeout)
-        my_roadblock.set_redis_server(redis_server)
-        my_roadblock.set_redis_password(args.roadblock_passwd)
-        my_roadblock.set_abort(abort)
-        my_roadblock.set_message_log(msgs_log_file)
-        my_roadblock.set_user_messages(messages)
-        if not wait_for is None:
-            my_roadblock.set_wait_for_cmd(wait_for)
-            my_roadblock.set_wait_for_log(wait_for_log)
-
-        rc = my_roadblock.run_it()
-        result_log_msg = "[%s] Roadblock resulted in return code %d" % (label, rc)
-        if rc != 0:
-            log.error(result_log_msg)
-        else:
-            log.info(result_log_msg)
-
-        stream = ""
-        with open(msgs_log_file, "r", encoding = "ascii") as msgs_log_file_fp:
-            for line in msgs_log_file_fp:
-                stream += line
-        log.info("[%s] Logged messages from roadblock:\n%s" % (label, stream))
-
-        if not wait_for is None:
-            stream = ""
-            with open(wait_for_log, "r", encoding = "ascii") as wait_for_log_fp:
-                for line in wait_for_log_fp:
-                    stream += line
-            log.info("[%s] Wait-for log from raodblock:\n%s" % (label, stream))
-
-    log.info("[%s] Total attempts: %d" % (label, attempts))
-    log.info("[%s] Returning %d" % (label, rc))
-    return rc
-
 def engine_init():
     """
     Construct messages to initialize the engines with metadata specific to them'
@@ -2302,137 +1686,16 @@ def engine_init():
                     }
                 }
 
-                env_vars_msgs.extend(create_roadblock_msg("follower", engine_name, "user-object", env_vars_payload))
+                env_vars_msgs.extend(endpoints.create_roadblock_msg("follower", engine_name, "user-object", env_vars_payload))
 
     env_vars_msg_file = settings["dirs"]["local"]["roadblock-msgs"] + "/env-vars.json"
     log.info("Writing follower env-vars messages to %s" % (env_vars_msg_file))
-    env_vars_msgs_json = dump_json(env_vars_msgs)
+    env_vars_msgs_json = endpoints.dump_json(env_vars_msgs)
     with open(env_vars_msg_file, "w", encoding = "ascii") as env_vars_msg_file_fp:
         env_vars_msg_file_fp.write(env_vars_msgs_json)
     log.info("Contents of %s:\n%s" % (env_vars_msg_file, env_vars_msgs_json))
 
     return env_vars_msg_file
-
-def prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, roadblock_name):
-    """
-    Prepare queued messages for distribution via roadblock
-
-    Args:
-        iteration_sample_dir (str): The directory where the iteration sample's files are stored
-        engine_tx_msgs_dir (str): Where to write messages to send
-        roadblock_name (str): The name of the roadblock that the messages should be sent for
-
-    Globals:
-        log: a logger instance
-
-    Returns:
-        user_msgs_file (str): The file containing the user messages if there are queued messages
-        or
-        None: If there are no queued messages
-    """
-    queued_msg_files = []
-    with os.scandir(engine_tx_msgs_dir) as tx_msgs_dir:
-        for entry in tx_msgs_dir:
-            queued_msg_files.append(entry.name)
-
-    if len(queued_msg_files) > 0:
-        log.info("Found queued messages in %s, preparing them to send" % (engine_tx_msgs_dir))
-
-        tx_sent_dir = engine_tx_msgs_dir + "-sent"
-        my_make_dirs(tx_sent_dir)
-
-        user_msgs = []
-        for msg_file in queued_msg_files:
-            msg_file_full_path = engine_tx_msgs_dir + "/" + msg_file
-            log.info("Importing %s" % (msg_file_full_path))
-            msg_file_json,err = load_json_file(msg_file_full_path)
-            if msg_file_json is None:
-                log.error("Failed to load user messages from %s with error '%s'" % (msg_file_full_path, err))
-            else:
-                log.info("Adding user messages from %s" % (msg_file_full_path))
-                user_msgs.extend(msg_file_json)
-
-                new_msg_file_full_path = tx_sent_dir + "/" + msg_file
-                log.info("Moving user message file from %s to %s" % (msg_file_full_path, new_msg_file_full_path))
-                os.replace(msg_file_full_path, new_msg_file_full_path)
-        payload = {
-            "sync": roadblock_name
-        }
-        user_msgs.extend(create_roadblock_msg("all", "all", "user-object", payload))
-
-        user_msgs_file = "%s/rb-msgs-%s.json" % (iteration_sample_dir, roadblock_name)
-        log.info("Writing user messages to %s" % (user_msgs_file))
-        user_msgs_file_json = dump_json(user_msgs)
-        with open(user_msgs_file, "w", encoding = "ascii") as user_msgs_file_fp:
-            user_msgs_file_fp.write(user_msgs_file_json)
-        log.info("Contents of %s:\n%s" % (user_msgs_file, user_msgs_file_json))
-
-        return user_msgs_file
-    else:
-        log.info("No queued messages found in %s" % (engine_tx_msgs_dir))
-        return None
-
-def evaluate_roadblock(quit, abort, roadblock_name, roadblock_rc, iteration_sample, engine_rx_msgs_dir):
-    """
-    Evaluate the status of a completed roadblock and it's affect on the test
-
-    Args:
-        quit (bool): Should the entire test be quit
-        abort (bool): Should the iteration be aborted
-        roadblock_name (str): The name of the roadblock being evaluated
-        iteration_sample (dict): The data structure representing the specific iteration sample being evaluated
-        engine_rx_msgs_dir (str): Where to look for received messages
-
-    Globals:
-        log: a logger instance
-        roadblock_exits (dict): A mapping of specific roadblock "events" to their associated return code
-
-    Returns:
-        abort (bool): The current abort value
-        quit (bool): The current quit value
-    """
-    if roadblock_rc != 0:
-        if roadblock_rc == roadblock_exits["timeout"]:
-            log.error("Roadblock '%s' timed out, attempting to exit and cleanly finish the run" % (roadblock_name))
-            quit = True
-        elif roadblock_rc == roadblock_exits["abort"]:
-            log.warning("Roadblock '%s' received an abort, stopping sample" % (roadblock_name))
-
-            iteration_sample["attempt-fail"] = 1
-            iteration_sample["failures"] += 1
-            log.info("iteration sample failures is now %d" % (iteration_sample["failures"]))
-
-            if iteration_sample["failures"] >= args.max_sample_failures:
-                iteration_sample["complete"] = True
-                log.error("A maximum of %d failures for iteration %d has been reached" % (iteration_sample["failures"], iteration_sample["iteration-id"]))
-
-            abort = True
-
-    msgs_log_file = engine_rx_msgs_dir + "/" + roadblock_name + ".json"
-    path = Path(msgs_log_file)
-    if path.exists() and path.is_file():
-        log.info("Found received messages file: %s" % (msgs_log_file))
-
-        split_roadblock_name = roadblock_name.split(":")
-        roadblock_label = split_roadblock_name[1]
-
-        msgs_json,err = load_json_file(msgs_log_file)
-        if not msgs_json is None:
-            if "received" in msgs_json:
-                counter = 0
-                for msg in msgs_json["received"]:
-                    if msg["payload"]["message"]["command"] == "user-object":
-                        counter += 1
-                        msg = "%s:%d" % (roadblock_label, counter)
-                        msg_outfile = engine_rx_msgs_dir + "/" + msg
-                        msg_outfile_json = dump_json(msg["paylod"]["message"]["user-object"])
-                        log.info("Found user-object message and saved it to %s:\nmessage:\n%s" % (msg_outfile, msg_outfile_json))
-                        with open(msg_outfile, "w", encoding = "ascii") as msg_outfile_fp:
-                            msg_outfile_fp.write(msg_outfile_json)
-        else:
-            log.error("Failed to load %s due to error '%s'" % (msgs_log_file, str(err)))
-
-    return quit,abort
 
 def test_start(msgs_dir, test_id, tx_msgs_dir):
     """
@@ -2506,17 +1769,17 @@ def test_start(msgs_dir, test_id, tx_msgs_dir):
                                     process_msg = True
 
                             if process_msg:
-                                log.info("Processing received message:\n%s" % (dump_json(msg["payload"])))
+                                log.info("Processing received message:\n%s" % (endpoints.dump_json(msg["payload"])))
 
                                 # punching a hole through the fireall would go here
 
                                 log.info("Creating a message to send to the client engine '%s' with IP and port info" % (client_engine))
-                                msg = create_roadblock_msg("follower", client_engine, "user-object", msg["payload"]["message"]["user-object"])
+                                msg = endpoints.create_roadblock_msg("follower", client_engine, "user-object", msg["payload"]["message"]["user-object"])
 
                                 msg_file = tx_msgs_dir + "/server-ip-" + server_engine + ".json"
                                 log.info("Writing follower service-ip message to '%s'" % (msg_file))
                                 with open(msg_file, "w", encoding = "ascii") as msg_file_fp:
-                                    msg_file_fp.write(dump_json(msg))
+                                    msg_file_fp.write(endpoints.dump_json(msg))
         else:
             log.error("Failed to load '%s' due to error '%s'" % (this_msg_file, str(err)))
     else:
@@ -2542,267 +1805,6 @@ def test_stop():
     log.info("...nothing to do here, run along...")
     log.info("Returning from test_stop()")
     return
-
-def process_bench_roadblocks():
-    """
-    Handle the running and evaluation of roadblocks while looping through the iterations and samples
-
-    Args:
-        None
-
-    Globals:
-        log: a logger instance
-        settings (dict): the one data structure to rule then all
-
-    Returns:
-        0
-    """
-    log.info("Starting to process benchmark roadblocks")
-
-    rc = do_roadblock(label = "setup-bench-begin",
-                      timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-    if rc != 0:
-        return rc
-
-    iteration_sample_data = []
-
-    log.info("Initializing data structures")
-    with open(settings["dirs"]["local"]["engine-cmds"] + "/client/1/start") as bench_cmds_fp:
-        for line in bench_cmds_fp:
-            split = line.split(" ")
-            iteration_sample = split[0]
-            split = iteration_sample.split("-")
-            iteration_id = int(split[0])
-            sample_id = int(split[1])
-
-            log.info("iteration_sample=%s iteration_id=%s sample_id=%s" % (iteration_sample, iteration_id, sample_id))
-
-            obj = {
-                "iteration-sample": iteration_sample,
-                "iteration-id": iteration_id,
-                "sample-id": sample_id,
-                "failures": 0,
-                "complete": False,
-                "attempt-num": 0,
-                "attempt-fail": 0
-            }
-            iteration_sample_data.append(obj)
-
-    log.info("Total tests: %d" % (len(iteration_sample_data)))
-
-    rc = do_roadblock(label = "setup-bench-end",
-                      timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-    if rc != 0:
-        return rc
-
-    quit = False
-    abort = False
-    timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"]
-    current_test = 0
-
-    for iteration_sample_idx,iteration_sample in enumerate(iteration_sample_data):
-        if quit:
-            break
-
-        current_test += 1
-
-        iteration_sample_dir = "%s/iteration-%d/sample-%d" % (settings["dirs"]["local"]["endpoint"], iteration_sample["iteration-id"], iteration_sample["sample-id"])
-        engine_msgs_dir = "%s/msgs" % (iteration_sample_dir)
-        engine_tx_msgs_dir = "%s/tx" % (engine_msgs_dir)
-        engine_rx_msgs_dir = "%s/rx" % (engine_msgs_dir)
-        log.info("Creating iteration+sample directories")
-        for current_dir in [ iteration_sample_dir, engine_msgs_dir, engine_tx_msgs_dir, engine_rx_msgs_dir ]:
-            my_make_dirs(current_dir)
-
-        abort = False
-        while not quit and not abort and not iteration_sample["complete"] and iteration_sample["failures"] < args.max_sample_failures:
-            iteration_sample["attempt-fail"] = 0
-            iteration_sample["attempt-num"] += 1
-
-            log.info("Starting iteration %d sample %d (test %d of %d) attempt number %d of %d" %
-                     (
-                         iteration_sample["iteration-id"],
-                         iteration_sample["sample-id"],
-                         current_test,
-                         len(iteration_sample_data),
-                         iteration_sample["attempt-num"],
-                         args.max_sample_failures
-                     ))
-
-            rb_name = None
-            test_id = "%d-%d-%d" % (iteration_sample["iteration-id"], iteration_sample["sample-id"], iteration_sample["attempt-num"])
-            rb_prefix = "%s:" % (test_id)
-
-            ####################################################################
-            rb_name = "%s%s" % (rb_prefix, "infra-start-begin")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-
-            rb_name = "%s%s" % (rb_prefix, "infra-start-end")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-            ####################################################################
-            rb_name = "%s%s" % (rb_prefix, "server-start-begin")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-
-            rb_name = "%s%s" % (rb_prefix, "server-start-end")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-            ####################################################################
-            rb_name = "%s%s" % (rb_prefix, "endpoint-start-begin")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-
-            test_start(settings["dirs"]["local"]["roadblock-msgs"], test_id, engine_tx_msgs_dir)
-
-            rb_name = "%s%s" % (rb_prefix, "endpoint-start-end")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-            ####################################################################
-            rb_name = "%s%s" % (rb_prefix, "client-start-begin")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-
-            msgs_log_file = settings["dirs"]["local"]["roadblock-msgs"] + "/" + rb_name + ".json"
-            path = Path(msgs_log_file)
-            if path.exists() and path.is_file():
-                msgs_json,err = load_json_file(msgs_log_file)
-                if not msgs_json is None:
-                    new_timeout = None
-                    if "received" in msgs_json:
-                        for msg in msgs_json["received"]:
-                            if msg["payload"]["message"]["command"] == "user-object" and "timeout" in msg["payload"]["message"]["user-object"]:
-                                new_timeout = msg["payload"]["message"]["user-object"]["timeout"]
-                                break
-                    if new_timeout is None:
-                        log.warning("Could not find new client-start-end timeout value")
-                    else:
-                        timeout = int(new_timeout)
-                        log.info("Found new client-start-end timeout value: %d" % (timeout))
-                else:
-                    log.error("Failed to load %s due to error '%s'" % (msgs_log_file, str(err)))
-            else:
-                log.warning("Could not find %s" % (msgs_log_file))
-
-            rb_name = "%s%s" % (rb_prefix, "client-start-end")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-            ####################################################################
-            if timeout != settings["rickshaw"]["roadblock"]["timeouts"]["default"]:
-                timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"]
-                log.info("Resetting timeout value: %s" % (timeout))
-
-            rb_name = "%s%s" % (rb_prefix, "client-stop-begin")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-
-            rb_name = "%s%s" % (rb_prefix, "client-stop-end")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-            ####################################################################
-            rb_name = "%s%s" % (rb_prefix, "endpoint-stop-begin")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-
-            rb_name = "%s%s" % (rb_prefix, "endpoint-stop-end")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-            ####################################################################
-            rb_name = "%s%s" % (rb_prefix, "server-stop-begin")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-
-            test_stop()
-
-            rb_name = "%s%s" % (rb_prefix, "server-stop-end")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-            ####################################################################
-            rb_name = "%s%s" % (rb_prefix, "infra-stop-begin")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-
-            rb_name = "%s%s" % (rb_prefix, "infra-stop-end")
-            user_msgs_file = prepare_roadblock_user_msgs_file(iteration_sample_dir, engine_tx_msgs_dir, rb_name)
-            rc = do_roadblock(label = rb_name,
-                              timeout = timeout,
-                              messages = user_msgs_file)
-            quit,abort = evaluate_roadblock(quit, abort, rb_name, rc, iteration_sample, engine_rx_msgs_dir)
-            ####################################################################
-
-            sample_result = None
-            if iteration_sample["attempt-fail"] == 0 and not abort and not quit:
-                iteration_sample["complete"] = True
-                sample_result = "successfully"
-            else:
-                sample_result = "unsuccessfully"
-
-                if abort:
-                    log.warning("An abort signal has been encountered for this sample")
-
-                if quit:
-                    log.error("A quit signal has been encountered")
-
-            log.info("Completed iteration %d sample %d (test %d of %d) attempt number %d of %d %s" %
-                     (
-                         iteration_sample["iteration-id"],
-                         iteration_sample["sample-id"],
-                         current_test,
-                         len(iteration_sample_data),
-                         iteration_sample["attempt-num"],
-                         args.max_sample_failures,
-                         sample_result
-                     ))
-
-    log.info("Final summary of iteration sample data:\n%s" % (dump_json(iteration_sample_data)))
-
-    return 0
 
 def collect_podman_log(thread_name, remote_name, engine_name, container_name, connection):
     """
@@ -2830,7 +1832,7 @@ def collect_podman_log(thread_name, remote_name, engine_name, container_name, co
     cmd_rc = 1
 
     while cmd_rc != 0 and cmd_attempt <= cmd_retries:
-        result = run_remote(connection, "podman logs --timestamps " + container_name + " | xz -c | base64")
+        result = endpoints.run_remote(connection, "podman logs --timestamps " + container_name + " | xz -c | base64")
         if result.exited != 0:
             thread_logger(thread_name, "Collecting podman log for %s failed with return code %d on attempt %d of %d:\nstdout:\n%sstderr:\n%s" %
                         (
@@ -2877,7 +1879,7 @@ def collect_chroot_log(thread_name, remote_name, engine_name, container_name, co
     thread_logger(thread_name, "Collecting chroot log", remote_name = remote_name, engine_name = engine_name)
 
     remote_log_file = settings["dirs"]["remote"]["logs"] + "/" + engine_name + ".txt"
-    result = run_remote(connection, "cat " + remote_log_file + " | xz -c | base64")
+    result = endpoints.run_remote(connection, "cat " + remote_log_file + " | xz -c | base64")
     if result.exited != 0:
         thread_logger(thread_name, "Collecting chroot log for %s failed with return code %d:\nstdout:\n%sstderr:\n%s" %
                       (
@@ -2893,14 +1895,14 @@ def collect_chroot_log(thread_name, remote_name, engine_name, container_name, co
             log_file_fp.write(base64.b64decode(result.stdout))
         thread_logger(thread_name, "Wrote chroot log to %s" % (log_file), remote_name = remote_name, engine_name = engine_name)
 
-        result = run_remote(connection, "rm --verbose " + remote_log_file)
+        result = endpoints.run_remote(connection, "rm --verbose " + remote_log_file)
         thread_logger(thread_name, "Removal of engine log for '%s' gave return code %d:\nstdout:\n%stderr:\n%s" %
                       (
                           engine_name,
                           result.exited,
                           result.stdout,
                           result.stderr
-                      ), log_level =  get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+                      ), log_level =  endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
         return True
 
 def remove_rickshaw_settings(connection, thread_name, remote_name, engine_name):
@@ -2920,8 +1922,8 @@ def remove_rickshaw_settings(connection, thread_name, remote_name, engine_name):
         None
     """
     remote_rickshaw_settings = settings["dirs"]["remote"]["data"] + "/rickshaw-settings.json.xz"
-    result = run_remote(connection, "if [ -e \"" + remote_rickshaw_settings + "\" ]; then rm --verbose \"" + remote_rickshaw_settings + "\"; else echo \"rickshaw settings already removed\"; fi")
-    thread_logger(thread_name, "Removal of rickshaw settings '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (remote_rickshaw_settings, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+    result = endpoints.run_remote(connection, "if [ -e \"" + remote_rickshaw_settings + "\" ]; then rm --verbose \"" + remote_rickshaw_settings + "\"; else echo \"rickshaw settings already removed\"; fi")
+    thread_logger(thread_name, "Removal of rickshaw settings '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (remote_rickshaw_settings, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
     return
 
@@ -2942,8 +1944,8 @@ def remove_ssh_private_key(connection, thread_name, remote_name, engine_name):
         None
     """
     remote_ssh_private_key = settings["dirs"]["remote"]["data"] + "/rickshaw_id.rsa"
-    result = run_remote(connection, "if [ -e \"" + remote_ssh_private_key + "\" ]; then rm --verbose \"" + remote_ssh_private_key + "\"; else echo \"ssh private key already removed\"; fi")
-    thread_logger(thread_name, "Removal of ssh private key '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (remote_ssh_private_key, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+    result = endpoints.run_remote(connection, "if [ -e \"" + remote_ssh_private_key + "\" ]; then rm --verbose \"" + remote_ssh_private_key + "\"; else echo \"ssh private key already removed\"; fi")
+    thread_logger(thread_name, "Removal of ssh private key '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (remote_ssh_private_key, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
     return
 
@@ -2966,12 +1968,12 @@ def destroy_podman(thread_name, remote_name, engine_name, container_name, connec
     """
     thread_logger(thread_name, "Destroying podman", remote_name = remote_name, engine_name = engine_name)
 
-    result = run_remote(connection, "podman rm --force " + container_name)
-    thread_logger(thread_name, "Removal of pod '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (engine_name, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+    result = endpoints.run_remote(connection, "podman rm --force " + container_name)
+    thread_logger(thread_name, "Removal of pod '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (engine_name, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
     remote_env_file = settings["dirs"]["remote"]["cfg"] + "/" + engine_name + "_env.txt"
-    result = run_remote(connection, "rm --verbose " + remote_env_file)
-    thread_logger(thread_name, "Removal of env file  '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (remote_env_file, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+    result = endpoints.run_remote(connection, "rm --verbose " + remote_env_file)
+    thread_logger(thread_name, "Removal of env file  '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (remote_env_file, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
     remove_ssh_private_key(connection, thread_name, remote_name, engine_name)
 
@@ -3000,63 +2002,21 @@ def destroy_chroot(thread_name, remote_name, engine_name, container_name, connec
     thread_logger(thread_name, "Destroying chroot", remote_name = remote_name, engine_name = engine_name)
 
     for mount in chroot_info["mounts"]["regular"]:
-        result = run_remote(connection, "umount --verbose " + mount)
-        thread_logger(thread_name, "regular unmounting of '%s' resulted in record code %d:\nstdout:\n%sstderr:\n%s" % (mount, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+        result = endpoints.run_remote(connection, "umount --verbose " + mount)
+        thread_logger(thread_name, "regular unmounting of '%s' resulted in record code %d:\nstdout:\n%sstderr:\n%s" % (mount, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
     for mount in chroot_info["mounts"]["rbind"]:
-        result = run_remote(connection, "umount --verbose --recursive " + mount)
-        thread_logger(thread_name, "recursive unmounting of '%s' resulted in record code %d:\nstdout:\n%sstderr:\n%s" % (mount, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+        result = endpoints.run_remote(connection, "umount --verbose --recursive " + mount)
+        thread_logger(thread_name, "recursive unmounting of '%s' resulted in record code %d:\nstdout:\n%sstderr:\n%s" % (mount, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
-    result = run_remote(connection, "podman rm --force " + container_name)
-    thread_logger(thread_name, "Removal of pod '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (engine_name, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
+    result = endpoints.run_remote(connection, "podman rm --force " + container_name)
+    thread_logger(thread_name, "Removal of pod '%s' resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (engine_name, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, engine_name = engine_name)
 
     remove_ssh_private_key(connection, thread_name, remote_name, engine_name)
 
     remove_rickshaw_settings(connection, thread_name, remote_name, engine_name)
 
     return
-
-def gmtimestamp_to_gmepoch(gmtimestamp):
-    """
-    Convert the provided formatted UTC timestamp to a UTC epoch timestamp
-
-    Args:
-        gmtimestamp (str): A formatted UTC tiemstamp such as "2024-04-11 14:30:39 +0000 UTC"
-
-    Globals:
-        None
-
-    Returns:
-        gmepoch (int): A UTC epoch timestamp such as 1712949457
-    """
-    gmepoch = None
-
-    time_struct = time.strptime(gmtimestamp, "%Y-%m-%d %H:%M:%S %z %Z")
-
-    gmepoch = calendar.timegm(time_struct)
-
-    return gmepoch
-
-def image_expiration_gmepoch():
-    """
-    Determine the UTC epoch timetamp that any image created before is expired
-
-    Args:
-        None
-
-    Globals:
-        None
-
-    Returns:
-        gmepoch (int): A UTC epoch timestamp such as 1712949457 from 2 weeks ago.  Any image with a creation
-                       date older than this is expired
-    """
-    #           seconds/min  minutes/hour  hours/day  days
-    two_weeks = 60           * 60          * 24       * 14
-
-    gmepoch = calendar.timegm(time.gmtime()) - two_weeks
-
-    return gmepoch
 
 def remove_image(thread_name, remote_name, log_prefix, connection, image):
     """
@@ -3076,8 +2036,8 @@ def remove_image(thread_name, remote_name, log_prefix, connection, image):
     """
     thread_logger(thread_name, "Removing image '%s'" % (image), remote_name = remote_name, log_prefix = log_prefix)
 
-    result = run_remote(connection, "podman rmi " + image)
-    thread_logger(thread_name, "Removing podman image '%s' gave return code %d:\nstdout:\n%sstderr:\n%s" % (image, result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, log_prefix = log_prefix)
+    result = endpoints.run_remote(connection, "podman rmi " + image)
+    thread_logger(thread_name, "Removing podman image '%s' gave return code %d:\nstdout:\n%sstderr:\n%s" % (image, result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, log_prefix = log_prefix)
 
     return
 
@@ -3102,14 +2062,14 @@ def remote_image_manager(thread_name, remote_name, connection, image_max_cache_s
     log_prefix = "RIM"
     thread_logger(thread_name, "Performing container image management", remote_name = remote_name, log_prefix = log_prefix)
 
-    result = run_remote(connection, "podman images --all")
+    result = endpoints.run_remote(connection, "podman images --all")
     thread_logger(thread_name, "All podman images on this remote host before running image manager:\nstdout:\n%sstderr:\n%s" % (result.stdout, result.stderr), remote_name = remote_name)
 
     images = dict()
     images["rickshaw"] = dict()
     images["podman"] = dict()
 
-    result = run_remote(connection, "cat /var/lib/crucible/remotehosts-container-image-census")
+    result = endpoints.run_remote(connection, "cat /var/lib/crucible/remotehosts-container-image-census")
     if result.exited != 0:
         thread_logger(thread_name, "Reading image census returned %d:\nstdout:\n%sstderr:\n%s" % (result.exited, result.stdout, result.stderr), log_level = "error", remote_name = remote_name, log_prefix = log_prefix)
         return
@@ -3130,9 +2090,9 @@ def remote_image_manager(thread_name, remote_name, connection, image_max_cache_s
 
         images["rickshaw"][image_name]["timestamps"].append(image_timestamp)
         images["rickshaw"][image_name]["run-ids"][image_timestamp] = image_run_id
-    thread_logger(thread_name, "images[rickshaw]:\n%s" % (dump_json(images["rickshaw"])), remote_name = remote_name, log_prefix = log_prefix)
+    thread_logger(thread_name, "images[rickshaw]:\n%s" % (endpoints.dump_json(images["rickshaw"])), remote_name = remote_name, log_prefix = log_prefix)
 
-    result = run_remote(connection, "podman images --format='{{.Repository}}:{{.Tag}}|{{.CreatedAt}}'")
+    result = endpoints.run_remote(connection, "podman images --format='{{.Repository}}:{{.Tag}}|{{.CreatedAt}}'")
     if result.exited != 0:
         thread_logger(thread_name, "Getting podman images returned %d:\nstdout:\n%sstderr:\n%s" % (result.exited, result.stdout, result.stderr), log_level = "error", remote_name = remote_name, log_prefix = log_prefix)
         return
@@ -3144,11 +2104,11 @@ def remote_image_manager(thread_name, remote_name, connection, image_max_cache_s
         if not image_name in images["podman"]:
             images["podman"][image_name] = {
                 "created-input": image_timestamp,
-                "created": gmtimestamp_to_gmepoch(image_timestamp)
+                "created": endpoints.gmtimestamp_to_gmepoch(image_timestamp)
             }
-    thread_logger(thread_name, "images[podman]:\n%s" % (dump_json(images["podman"])), remote_name = remote_name, log_prefix = log_prefix)
+    thread_logger(thread_name, "images[podman]:\n%s" % (endpoints.dump_json(images["podman"])), remote_name = remote_name, log_prefix = log_prefix)
 
-    image_expiration = image_expiration_gmepoch()
+    image_expiration = endpoints.image_expiration_gmepoch()
     thread_logger(thread_name, "Images created before %d will be considered expired" % (image_expiration), remote_name = remote_name, log_prefix = log_prefix)
 
     deletes = []
@@ -3167,7 +2127,7 @@ def remote_image_manager(thread_name, remote_name, connection, image_max_cache_s
         del images["rickshaw"][image]["timestamps"]
     for image in deletes:
         del images["rickshaw"][image]
-    thread_logger(thread_name, "images[rickshaw]:\n%s" % (dump_json(images["rickshaw"])), remote_name = remote_name, log_prefix = log_prefix)
+    thread_logger(thread_name, "images[rickshaw]:\n%s" % (endpoints.dump_json(images["rickshaw"])), remote_name = remote_name, log_prefix = log_prefix)
 
     deletes = dict()
     deletes["rickshaw"] = []
@@ -3193,15 +2153,15 @@ def remote_image_manager(thread_name, remote_name, connection, image_max_cache_s
     for kind in deletes.keys():
         for image in deletes[kind]:
             del images[kind][image]
-    thread_logger(thread_name, "images[rickshaw]:\n%s" % (dump_json(images["rickshaw"])), remote_name = remote_name, log_prefix = log_prefix)
-    thread_logger(thread_name, "images[podman]:\n%s" % (dump_json(images["podman"])), remote_name = remote_name, log_prefix = log_prefix)
+    thread_logger(thread_name, "images[rickshaw]:\n%s" % (endpoints.dump_json(images["rickshaw"])), remote_name = remote_name, log_prefix = log_prefix)
+    thread_logger(thread_name, "images[podman]:\n%s" % (endpoints.dump_json(images["podman"])), remote_name = remote_name, log_prefix = log_prefix)
 
     cache_size = 0
 
     if cache_size < image_max_cache_size:
         thread_logger(thread_name, "Attemping to cache the most recently used image(s)", remote_name = remote_name, log_prefix = log_prefix)
         sorted_images = sorted(images["rickshaw"].items(), key = lambda x: (x[1]["latest-usage"], x[0]), reverse = True)
-        thread_logger(thread_name, "latest sorted images[rickshaw]:\n%s" % (dump_json(sorted_images)), remote_name = remote_name, log_prefix = log_prefix)
+        thread_logger(thread_name, "latest sorted images[rickshaw]:\n%s" % (endpoints.dump_json(sorted_images)), remote_name = remote_name, log_prefix = log_prefix)
 
         image = sorted_images[0][0]
         image_last_used = sorted_images[0][1]["latest-usage"]
@@ -3230,14 +2190,14 @@ def remote_image_manager(thread_name, remote_name, connection, image_max_cache_s
         else:
             thread_logger(thread_name, "Image cache is full", remote_name = remote_name, log_prefix = log_prefix)
 
-        thread_logger(thread_name, "images[rickshaw]:\n%s" % (dump_json(images["rickshaw"])), remote_name = remote_name, log_prefix = log_prefix)
+        thread_logger(thread_name, "images[rickshaw]:\n%s" % (endpoints.dump_json(images["rickshaw"])), remote_name = remote_name, log_prefix = log_prefix)
     else:
         thread_logger(thread_name, "No images can be cached due to maximum image cache size of %d" % (image_max_cache_size), remote_name = remote_name, log_prefix = log_prefix)
 
     if cache_size < image_max_cache_size:
         thread_logger(thread_name, "Attempting to cache the most used image(s)", remote_name = remote_name, log_prefix = log_prefix)
         sorted_images = sorted(images["rickshaw"].items(), key = lambda x: (x[1]["uses"], x[0]), reverse = True)
-        thread_logger(thread_name, "usage sorted images[rickshaw]:\n%s" % (dump_json(sorted_images)), remote_name = remote_name, log_prefix = log_prefix)
+        thread_logger(thread_name, "usage sorted images[rickshaw]:\n%s" % (endpoints.dump_json(sorted_images)), remote_name = remote_name, log_prefix = log_prefix)
 
         loop = True
         while loop and cache_size < image_max_cache_size:
@@ -3281,7 +2241,7 @@ def remote_image_manager(thread_name, remote_name, connection, image_max_cache_s
                 thread_logger(thread_name, "All rickshaw images are cached", remote_name = remote_name, log_prefix = log_prefix)
                 break
 
-        thread_logger(thread_name, "images[rickshaw]:\n%s" % (dump_json(images["rickshaw"])), remote_name = remote_name, log_prefix = log_prefix)
+        thread_logger(thread_name, "images[rickshaw]:\n%s" % (endpoints.dump_json(images["rickshaw"])), remote_name = remote_name, log_prefix = log_prefix)
     else:
         if cache_size == 0:
             thread_logger(thread_name, "No images can be cached due to maximum image cache size of %d" % (image_max_cache_size), remote_name = remote_name, log_prefix = log_prefix)
@@ -3297,10 +2257,10 @@ def remote_image_manager(thread_name, remote_name, connection, image_max_cache_s
     if all_cached:
         thread_logger(thread_name, "All rickshaw images are cached", remote_name = remote_name, log_prefix = log_prefix)
 
-    result = run_remote(connection, "podman image prune -f")
-    thread_logger(thread_name, "Pruning dangling images resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote_name, log_prefix = log_prefix)
+    result = endpoints.run_remote(connection, "podman image prune -f")
+    thread_logger(thread_name, "Pruning dangling images resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote_name, log_prefix = log_prefix)
 
-    result = run_remote(connection, "podman images --all")
+    result = endpoints.run_remote(connection, "podman images --all")
     thread_logger(thread_name, "All podman images on this remote host after running image manager:\nstdout:\n%sstderr:\n%s" % (result.stdout, result.stderr), remote_name = remote_name)
 
     return
@@ -3347,14 +2307,14 @@ def shutdown_engines_worker_thread(thread_id, work_queue, threads_rcs):
         thread_logger(thread_name, "Processing remote '%s' at index %d" % (remote["config"]["host"], remote_idx), remote_name = remote_name)
         thread_logger(thread_name, "Remote user is %s" % (remote["config"]["settings"]["remote-user"]), remote_name = remote_name)
 
-        with remote_connection(remote["config"]["host"], remote["config"]["settings"]["remote-user"]) as con:
-            result = run_remote(con, "mount")
+        with endpoints.remote_connection(remote["config"]["host"], remote["config"]["settings"]["remote-user"]) as con:
+            result = endpoints.run_remote(con, "mount")
             thread_logger(thread_name, "All mounts on this remote host:\nstdout:\n%sstderr:\n%s" % (result.stdout, result.stderr), remote_name = remote_name)
 
-            result = run_remote(con, "podman ps --all")
+            result = endpoints.run_remote(con, "podman ps --all")
             thread_logger(thread_name, "All podman pods on this remote host:\nstdout:\n%sstderr:\n%s" % (result.stdout, result.stderr), remote_name = remote_name)
 
-            result = run_remote(con, "podman mount")
+            result = endpoints.run_remote(con, "podman mount")
             thread_logger(thread_name, "All podman container mounts on this remote host:\nstdout:\n%sstderr:\n%s" % (result.stdout, result.stderr), remote_name = remote_name)
 
             for engine in remote["engines"]:
@@ -3381,13 +2341,13 @@ def shutdown_engines_worker_thread(thread_id, work_queue, threads_rcs):
                             if success:
                                 destroy_chroot(thread_name, remote_name, engine_name, container_name, con, remote["chroots"][engine_name])
 
-            result = run_remote(con, "mount")
+            result = endpoints.run_remote(con, "mount")
             thread_logger(thread_name, "All mounts on this remote host:\nstdout:\n%sstderr:\n%s" % (result.stdout, result.stderr), remote_name = remote_name)
 
-            result = run_remote(con, "podman ps --all")
+            result = endpoints.run_remote(con, "podman ps --all")
             thread_logger(thread_name, "All podman pods on this remote host:\nstdout:\n%sstderr:\n%s" % (result.stdout, result.stderr), remote_name = remote_name)
 
-            result = run_remote(con, "podman mount")
+            result = endpoints.run_remote(con, "podman mount")
             thread_logger(thread_name, "All podman container mounts on this remote host:\nstdout:\n%sstderr:\n%s" % (result.stdout, result.stderr), remote_name = remote_name)
 
         thread_logger(thread_name, "Notifying work queue that job processing is complete", remote_name = remote_name)
@@ -3463,7 +2423,7 @@ def image_mgmt_worker_thread(thread_id, work_queue, threads_rcs):
 
         thread_logger(thread_name, "Remote user is %s" % (my_run_file_remote["config"]["settings"]["remote-user"]), remote_name = remote)
 
-        with remote_connection(remote, my_run_file_remote["config"]["settings"]["remote-user"]) as con:
+        with endpoints.remote_connection(remote, my_run_file_remote["config"]["settings"]["remote-user"]) as con:
             remote_image_manager(thread_name, remote, con, my_run_file_remote["config"]["settings"]["image-cache-size"])
 
         thread_logger(thread_name, "Notifying work queue that job processing is complete", remote_name = remote)
@@ -3557,35 +2517,35 @@ def collect_sysinfo_worker_thread(thread_id, work_queue, threads_rcs):
 
         thread_logger(thread_name, "Remote user is %s" % (my_run_file_remote["config"]["settings"]["remote-user"]), remote_name = remote)
 
-        with remote_connection(remote, my_run_file_remote["config"]["settings"]["remote-user"]) as con:
+        with endpoints.remote_connection(remote, my_run_file_remote["config"]["settings"]["remote-user"]) as con:
             local_dir = settings["dirs"]["local"]["sysinfo"] + "/" + remote
-            my_make_dirs(local_dir)
+            endpoints.my_make_dirs(local_dir)
 
             local_packrat_file = args.packrat_dir + "/packrat"
             remote_packrat_file = settings["dirs"]["remote"]["run"] + "/packrat"
             result = con.put(local_packrat_file, remote_packrat_file)
             thread_logger(thread_name, "Copied %s to %s:%s" % (local_packrat_file, remote, remote_packrat_file), remote_name = remote)
 
-            result = run_remote(con, remote_packrat_file + " " + settings["dirs"]["remote"]["sysinfo"])
-            thread_logger(thread_name, "Running packrat resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote)
+            result = endpoints.run_remote(con, remote_packrat_file + " " + settings["dirs"]["remote"]["sysinfo"])
+            thread_logger(thread_name, "Running packrat resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote)
 
-            result = run_remote(con, "tar --create --directory " + settings["dirs"]["remote"]["sysinfo"]  + " packrat-archive | xz --stdout | base64")
-            thread_logger(thread_name, "Transferring packrat files resulted in return code %d:\nstderr:\n%s" % (result.exited, result.stderr), log_level = get_result_log_level(result), remote_name = remote)
+            result = endpoints.run_remote(con, "tar --create --directory " + settings["dirs"]["remote"]["sysinfo"]  + " packrat-archive | xz --stdout | base64")
+            thread_logger(thread_name, "Transferring packrat files resulted in return code %d:\nstderr:\n%s" % (result.exited, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote)
 
             archive_file = local_dir + "/packrat.tar.xz"
             with open(archive_file, "wb") as archive_file_fp:
                 archive_file_fp.write(base64.b64decode(result.stdout))
             thread_logger(thread_name, "Wrote packrat archive to '%s'" % (archive_file), remote_name = remote)
 
-            result = run_local("xz --decompress --stdout " + archive_file + " | tar --extract --verbose --directory " + local_dir)
-            thread_logger(thread_name, "Unpacking packrat archive resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote)
+            result = endpoints.run_local("xz --decompress --stdout " + archive_file + " | tar --extract --verbose --directory " + local_dir)
+            thread_logger(thread_name, "Unpacking packrat archive resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote)
 
             path = Path(archive_file)
             path.unlink()
             thread_logger(thread_name, "Removed packrat archive '%s'" % (archive_file), remote_name = remote)
 
-            result = run_remote(con, "rm --recursive --force " + remote_packrat_file + " " + settings["dirs"]["remote"]["sysinfo"] + "/packrat-archive")
-            thread_logger(thread_name, "Removing remote packrat files resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (result.exited, result.stdout, result.stderr), log_level = get_result_log_level(result), remote_name = remote)
+            result = endpoints.run_remote(con, "rm --recursive --force " + remote_packrat_file + " " + settings["dirs"]["remote"]["sysinfo"] + "/packrat-archive")
+            thread_logger(thread_name, "Removing remote packrat files resulted in return code %d:\nstdout:\n%sstderr:\n%s" % (result.exited, result.stdout, result.stderr), log_level = endpoints.get_result_log_level(result), remote_name = remote)
 
         thread_logger(thread_name, "Notifying work queue that job processing is complete", remote_name = remote)
         work_queue.task_done()
@@ -3617,125 +2577,6 @@ def collect_sysinfo():
     create_thread_pool("Collect Sysinfo Worker Threads", "CSWT", sysinfo_work, worker_threads_count, collect_sysinfo_worker_thread)
 
     return 0
-
-def process_roadblocks():
-    """
-    Process the beginning and ending roadblocks associated with synchronizing a test
-
-    Args:
-        None
-
-    Globals:
-        args (namespace): the script's CLI parameters
-        settings (dict): the one data structure to rule then all
-
-    Returns:
-        0
-    """
-    log.info("Starting to process roadblocks")
-
-    new_followers_msg_payload = {
-        "new-followers": settings["engines"]["new-followers"]
-    }
-    new_followers_msg = create_roadblock_msg("all", "all", "user-object", new_followers_msg_payload)
-
-    new_followers_msg_file = settings["dirs"]["local"]["roadblock-msgs"] + "/new-followers.json"
-    log.info("Writing new followers message to %s" % (new_followers_msg_file))
-    with open(new_followers_msg_file, "w", encoding = "ascii") as new_followers_msg_file_fp:
-        new_followers_msg_file_fp.write(dump_json(new_followers_msg))
-
-    rc = do_roadblock(label = "endpoint-deploy-begin",
-                      timeout = args.endpoint_deploy_timeout,
-                      messages = new_followers_msg_file)
-    if rc != 0:
-        return rc
-    rc = do_roadblock(label = "endpoint-deploy-end",
-                      timeout = args.endpoint_deploy_timeout)
-    if rc != 0:
-        return rc
-
-    rc = do_roadblock(label = "engine-init-begin",
-                      timeout = settings["rickshaw"]["roadblock"]["timeouts"]["engine-start"])
-    if rc != 0:
-        return rc
-    engine_init_msgs = engine_init()
-    rc = do_roadblock(label = "engine-init-end",
-                      timeout = settings["rickshaw"]["roadblock"]["timeouts"]["engine-start"],
-                      messages = engine_init_msgs)
-    if rc != 0:
-        return rc
-
-    rc = do_roadblock(label = "get-data-begin",
-                      timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-    if rc != 0:
-        return rc
-    rc = do_roadblock(label = "get-data-end",
-                      timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-    if rc != 0:
-        return rc
-
-    rc = do_roadblock(label = "collect-sysinfo-begin",
-                      timeout = settings["rickshaw"]["roadblock"]["timeouts"]["collect-sysinfo"])
-    if rc != 0:
-        return rc
-    collect_sysinfo()
-    rc = do_roadblock(label = "collect-sysinfo-end",
-                      timeout = settings["rickshaw"]["roadblock"]["timeouts"]["collect-sysinfo"])
-    if rc != 0:
-        return rc
-
-    rc = do_roadblock(label = "start-tools-begin",
-                      timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-    if rc != 0:
-        return rc
-    rc = do_roadblock(label = "start-tools-end",
-                      timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-    if rc != 0:
-        return rc
-
-    rc = process_bench_roadblocks()
-    if rc != 0:
-        return rc
-
-    do_roadblock(label = "stop-tools-begin",
-                 timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-    do_roadblock(label = "stop-tools-end",
-                 timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-
-    do_roadblock(label = "send-data-begin",
-                 timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-    do_roadblock(label = "send-data-end",
-                 timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-
-    do_roadblock(label = "endpoint-cleanup-begin",
-                 timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-    remote_cleanup()
-    do_roadblock(label = "endpoint-cleanup-end",
-                 timeout = settings["rickshaw"]["roadblock"]["timeouts"]["default"])
-
-    return 0
-
-def setup_logger():
-    """
-    Setup the logging infrastructure that is used for everything except validation
-
-    Args:
-        None
-
-    Globals:
-        args (namespace): the script's CLI parameters
-
-    Returns:
-        a logging instance
-    """
-    log_format = '[LOG %(asctime)s %(levelname)s %(module)s %(funcName)s:%(lineno)d] %(message)s'
-    match args.log_level:
-        case "debug":
-            logging.basicConfig(level = logging.DEBUG, format = log_format, stream = sys.stdout)
-        case "normal" | _:
-            logging.basicConfig(level = logging.INFO, format = log_format, stream = sys.stdout)
-
-    return logging.getLogger(__file__)
 
 def thread_exception_hook(args):
     """
@@ -3781,9 +2622,9 @@ def main():
     if args.validate:
         return(validate())
 
-    log = setup_logger()
+    log = endpoints.setup_logger(args.log_level)
 
-    log_cli()
+    endpoints.log_cli(args)
     init_settings()
     if load_settings() != 0:
         return 1
@@ -3796,7 +2637,26 @@ def main():
     remotes_pull_images()
     set_total_cpu_partitions()
     launch_engines()
-    rc = process_roadblocks()
+
+    remotehosts_callbacks = {
+        "engine-init": engine_init,
+        "collect-sysinfo": collect_sysinfo,
+        "test-start": test_start,
+        "test-stop": test_stop,
+        "remote-cleanup": remote_cleanup
+    }
+    rc = endpoints.process_roadblocks(callbacks = remotehosts_callbacks,
+                                      roadblock_id = args.roadblock_id,
+                                      endpoint_label = args.endpoint_label,
+                                      endpoint_deploy_timeout = args.endpoint_deploy_timeout,
+                                      max_roadblock_attempts = args.max_rb_attempts,
+                                      roadblock_password = args.roadblock_passwd,
+                                      new_followers = settings["engines"]["new-followers"],
+                                      roadblock_messages_dir = settings["dirs"]["local"]["roadblock-msgs"],
+                                      roadblock_timeouts = settings["rickshaw"]["roadblock"]["timeouts"],
+                                      max_sample_failures = args.max_sample_failures,
+                                      engine_commands_dir = settings["dirs"]["local"]["engine-cmds"],
+                                      endpoint_dir = settings["dirs"]["local"]["endpoint"])
 
     log.info("Logging 'final' settings data structure")
     log_settings(mode = "settings")
